% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callffmpeg.R
\name{callffmpeg}
\alias{callffmpeg}
\title{Call FFmpeg with provided command}
\usage{
callffmpeg(cmd)
}
\arguments{
\item{cmd}{A string. The command passed to FFmpeg.}
}
\description{
\code{callffmpeg()} performs actual call for FFmpeg software
using \code{system()}.
It also generates some console outputs,
asking the users to wait for a second,
or telling them about the error during \code{system()} call.
}
\keyword{utilities}
