% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame2time.R
\name{frame2time}
\alias{frame2time}
\title{Convert frame number to timecode}
\usage{
frame2time(i, framerate, totalframes)
}
\arguments{
\item{i}{An integer. The index of targeted video frame.}

\item{framerate}{An integer. The frame rate of the video.}

\item{totalframes}{An integer. The total number of frames
in the video.}
}
\description{
\code{frame2time()} converts index of video frame (frame number)
to timecode expression like hh:mm:ss.ms
based on the designated frame rate.
If the index exceeds the existing frame,
it returns the initial or final frame index instead.
}
\keyword{internal}
\keyword{utilities}
