% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv-package.R
\docType{package}
\name{hsv-package}
\alias{hsv}
\alias{hsv-package}
\title{The R hsv package}
\description{
The hsv package offers easy front-end to handle avi video files through FFmpeg software. The functions were originally created for personal uses of the author and (mainly) my colleagues to analyze high-speed videos recorded in our experiments. (Thus this package has nothing to do with Hue-Saturation-Lightness color space, nor Herpes Simplex Virus.) Of course, the package can also be used in handling normal videos. Although the package is still under the construction and testing, it is already usable and thus the author decided to publish it open. Since hsv package is correctly a mere front-end to ffmpeg.exe, installation of FFmpeg software is mandatory before using this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/keimochizuki/hsv}
  \item Report bugs at \url{https://github.com/keimochizuki/hsv/issues}
}

}
\author{
\strong{Maintainer}: Kei Mochizuki \email{k31mczk-git@yahoo.co.jp}

}
