% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvCallFFmpeg.R
\name{hsvCallFFmpeg}
\alias{hsvCallFFmpeg}
\title{Call FFmpeg with arbitrary command}
\usage{
hsvCallFFmpeg(cmd)
}
\arguments{
\item{cmd}{A string. The command passed to FFmpeg.}
}
\description{
\code{\link[=hsvCallFFmpeg]{hsvCallFFmpeg()}} performs actual call for FFmpeg software
using \code{\link[base:system]{base::system()}}.
It also generates some console outputs,
asking the users to wait for a second,
or telling them about errors during system call.
This function can be used to perform arbitrary FFmpeg call
using complex filter settings or any other additional parameters
which are not currently supported in hsv package.
}
\keyword{utilities}
