% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvCaptureFrame.R
\name{hsvCaptureFrame}
\alias{hsvCaptureFrame}
\title{Extract video frames as png images}
\usage{
hsvCaptureFrame(
  infiles,
  frames,
  suffix = paste("\%0", floor(log10(max(frames))) + 1, "d", sep = ""),
  keepinfiles = TRUE,
  savedir = "."
)
}
\arguments{
\item{infiles}{Strings. The names of the avi files you want to screenshot.}

\item{frames}{Integers. The index of the frames to extract.}

\item{suffix}{A string. The suffix for the names of png files
concatenated to the names of input movie files.
Should be a format string for frame numbers to enable multi-frame export.
(e.g., "\%05d" for zero-padded five-digit frame indices.)
Otherwise, export of all png files will be done with
the same filename, overwriting the previous image repeatedly.}

\item{keepinfiles}{A logical. Whether to keep the input files
after the conversion.}

\item{savedir}{A string. The path to the directory
you want to save the output file(s).}
}
\value{
A list of strings. The names of the created png files.
}
\description{
\code{\link[=hsvCaptureFrame]{hsvCaptureFrame()}} extracts designated frames of the provided
movie file as png still images.
}
\examples{
\dontrun{
hsvCaptureFrame("video.avi", frames = seq(1, 1000, by = 50))
}

}
\keyword{utilities}
