% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvClipAvi.R
\name{hsvClipAvi}
\alias{hsvClipAvi}
\title{Extractor of clipped vedio from an avi file}
\usage{
hsvClipAvi(infile, outfile = "output.avi", from, to)
}
\arguments{
\item{infile}{A string. The name of the avi file you want to clip.}

\item{outfile}{A string. The name of the output avi file.}

\item{from}{An integer. The index of the frame in the input file
from where the clipped video starts.}

\item{to}{An integer. The index of the frame in the input file
to where the video should be clipped.}
}
\value{
A string. The names of the created avi file.
}
\description{
Extracts designated range of video and create a new avi file
from designated input source avi.
}
\details{
This function clips designated range in the input avi
into a new file, without performing any other video conversion.
It can be used, for example, to cut off certain range of the
video file to where your further analyses are applied.
Since no conversion is performed (with \verb{-c copy} option in \code{ffmpeg}),
this function normally takes not so long.
}
\examples{
\dontrun{
hsvClipAvi("input.avi", from = 1000, to = 1999)
}

}
\keyword{utilities}
