% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvGetMp4Header.R
\name{hsvGetMp4Header}
\alias{hsvGetMp4Header}
\title{Read out mp4 header}
\usage{
hsvGetMp4Header(infile, verbose = FALSE)
}
\arguments{
\item{infile}{A string. The name of the avi file you want to
read header information.}

\item{verbose}{A logical. Whether to print the extracted values
on the console lines.}
}
\value{
A list. Extracted header information.
}
\description{
\code{\link[=hsvGetMp4Header]{hsvGetMp4Header()}} extracts header information of
the first video stream in an mp4 file.
}
\details{
Unlike \code{\link[=hsvGetAviHeader]{hsvGetAviHeader()}}, this function reads header
information of a given mp4 file using 'ffprobe' utility.
}
\examples{
\dontrun{
hsvGetMp4Header("video.mp4", verbose = TRUE)
}

}
\keyword{utilities}
