% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvMakeExampleAvi.R
\name{hsvMakeExampleAvi}
\alias{hsvMakeExampleAvi}
\title{Make example avi files}
\usage{
hsvMakeExampleAvi(savedir = ".")
}
\arguments{
\item{savedir}{A string. The path to the directory
you want to save the output files.}
}
\value{
Strings. The names of the created avi files.
}
\description{
\code{\link[=hsvMakeExampleAvi]{hsvMakeExampleAvi()}} creates a set of two
high-speed videos in avi format.
These files can be used for your test purpose
in writing script with hsv package.
This may be handy because you do not need to
worry about these test files.
There are no worries about accidental file crash,
leakage of your experimental results,
intrusion of other people's privacy, or whatever.
}
\examples{
\dontrun{
hsvMakeExampleAvi()
}

}
\keyword{utilities}
