% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvStackAvi.R
\name{hsvStackAvi}
\alias{hsvStackAvi}
\title{Juxtapose multiple videos}
\usage{
hsvStackAvi(
  infiles,
  horizontal = TRUE,
  padding = 0,
  col = "#000000",
  crf = 20,
  outfile = "",
  suffix = "stacked",
  keepinfiles = TRUE,
  savedir = "."
)
}
\arguments{
\item{infiles}{Strings. The names of the avi files you want to merge.}

\item{horizontal}{A logical. Whether to stack the videos horizontally.
If set FALSE, videos are vertically stacked.}

\item{padding}{Integers of length \code{length(infiles) - 1}.
The sizes of paddings inserted between videos.}

\item{col}{A string. The color of the padding regions.
This can be provided either as the name of the color or
in hexadecimal form (like \verb{#000000}).}

\item{crf}{An integer. The constant rate factor (crf),
i.e., a value to determine the quality of the converted video
in FFmpeg (ranging from 0 to 51).
Smaller value indicate high quality video.
The default value for original \code{ffmpeg} command is 23.
Therefore, the default for this function is quality-oriented,
in exchange for possible larger file size.}

\item{outfile}{A string. The name of the output avi file.
If provided, this parameter override \code{suffix} argument
that is normally used to create output file names
in many other hsv functions.}

\item{suffix}{A string. The suffix to the file names
concatenated to the names of input avi files
to make those for the outputs.
Since this function creates one output from multiple inputs
(unlike other hsv functions),
only the name of the first input file is inherited
with the designated suffix.}

\item{keepinfiles}{A logical. Whether to keep the input files
after the stacking.}

\item{savedir}{A string. The path to the directory
you want to save the output file.}
}
\value{
A string. The name of the created avi file.
}
\description{
\code{\link[=hsvStackAvi]{hsvStackAvi()}} concatenates input avi files horizontally or vertically
to create a new, merged vedio with a juxtaposed view.
}
\details{
This function creates a stacked view from multiple input avi videos.
It is best used for the purpose to get a overview of the videos
recorded simultaneously with multiple cameras.
The input files should naturally have the same lengths.
The direction of stacking can be either horizontal (default)
or vertical.
Also, some additional conversion settings can be done
through optional arguments.
}
\examples{
\dontrun{
hsvStackAvi(c("video1.avi", "video2.avi"))
}

}
\keyword{utilities}
