% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinfiles.R
\name{checkinfiles}
\alias{checkinfiles}
\title{Check compatibility of source videos}
\usage{
checkinfiles(f, ext = "\\\\.avi$")
}
\arguments{
\item{f}{Strings. The names of the video files.}

\item{ext}{A string. A regular expression pattern for
file extensions allowed.}
}
\description{
\code{\link[=checkinfiles]{checkinfiles()}} performs routine checks of input
video files for many hsv functions.
These include confirmation of existence of the files,
as well as checking their file extentions.
}
\keyword{internal}
\keyword{utilities}
