% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksavedir.R
\name{checksavedir}
\alias{checksavedir}
\title{Check (and create if necessary) designated directory}
\usage{
checksavedir(d)
}
\arguments{
\item{d}{A string. The name of the directory.}
}
\description{
\code{checksavedir()} checks for the existence of a directory
where output files from hsv functions are saved.
If such a directory does not exist, new one is created.
}
\keyword{utilities}
