% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvCheckFFmpeg.R
\name{hsvCheckFFmpeg}
\alias{hsvCheckFFmpeg}
\title{Check FFmpeg availability}
\usage{
hsvCheckFFmpeg(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical. Whether to print the standard output
strings on R.}
}
\value{
A logical. Whether \code{ffmpeg} can be correctly called
on the system console.
This value is actually always \code{TRUE}, since this function
emit an error and stop the operation when \code{ffmepg} is
not available (instead of returning \code{FALSE}).
}
\description{
\code{\link[=hsvCheckFFmpeg]{hsvCheckFFmpeg()}} checks whether \code{ffmpeg} command is
available on the system.
}
\details{
Since video editing of hsv package is actually performed
by \code{ffmpeg.exe}, installation of FFmpeg is mandatory.
Users need to not only install FFmpeg executables
but also modify environment variables appropriately,
so that \code{ffmpeg} command can be correctly called
on the system's console.
This function checks the availability of \code{ffmpeg} command
by trying to ask version information to it.

If \code{ffmpeg} is not available, this function emits an error,
halting any subsequent operations.
Please check your system again to confirm FFmpeg installation.
If you have any difficuties in installation process,
refer to the README.md page of hsv package
available on the GitHub repository of the author.
}
\examples{
hsvCheckFFmpeg()

}
\keyword{utilities}
