% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvCropAvi.R
\name{hsvCropAvi}
\alias{hsvCropAvi}
\title{Crop a region of video area}
\usage{
hsvCropAvi(
  infiles,
  w,
  h,
  xoffset = 0,
  yoffset = 0,
  suffix = paste(w, h, sep = "x"),
  keepinfiles = TRUE,
  savedir = "."
)
}
\arguments{
\item{infiles}{Strings. The names of the avi files you want to crop.}

\item{w}{An integer. The width of the cropping region in pixels.}

\item{h}{An integer. The height of the cropping region in pixels.}

\item{xoffset}{An integer. The number of pixels skipped
toward x-axis before cropping region starts.
This value equals to the left margin omitted
from original videos.}

\item{yoffset}{An integer. The number of pixels skipped
toward y-axis before cropping region starts.
Since the y-axis starts from the upper end in image processing,
this value equals to the top (instead of bottom)
margin omitted from original videos.}

\item{suffix}{A string. The suffix to the file names
concatenated to the names of input avi files
to make those for the outputs.}

\item{keepinfiles}{A logical. Whether to keep the input files
after the conversion.}

\item{savedir}{A string. The path to the directory
you want to save the output file(s)}
}
\value{
A string. The names of the created avi file.
}
\description{
\code{hsvCropAvi()} extracts designated spatial region of videos
and create new avi files from input sources.
}
\details{
This function crops designated area in the input avi files.
without performing any other video conversion.
Such process is sometimes referred to as trimming,
especially in image processing.
However, the same word also means temporal clipping
in the field of video editing.
To avoid confusion, therefore, this process should be
better referred to as "area cropping" or "spatial trimming".
This functionality is used, for example,
when the field of view of your video recording is too wider
in relation to the targeted object.

The area of the cropped region is determined by
its width and height.
Also, the location of the region can be modified by
x (horizontal) and y (vertical) offsets.
Here caution should be exercised that,
as a tradition of image processing,
the origin of an image normally exists at the \emph{top}-left corner.
Therefore, the \code{yoffset} parameter determines
the \emph{top} margin (instead of bottom) to the cropped region.
}
\examples{
\dontrun{
hsvCropAvi("input.avi", w = 320, h = 240)
}

}
\keyword{utilities}
