% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvGetAviHeader.R
\name{hsvGetAviHeader}
\alias{hsvGetAviHeader}
\title{Checker of avi header information}
\usage{
hsvGetAviHeader(infile, verbose = FALSE)
}
\arguments{
\item{infile}{A string. The name of the avi file you want to clip.}

\item{verbose}{A logical. Whether to print the extracted values
on the console lines.}
}
\value{
A list. Extracted header information.
}
\description{
Extracts header information of the first video stream
in an avi file.
}
\details{
An avi file starts with a header data with a predetermined
format, followed by the actual video data field.
A lot of useful meta information can be obtained from avi header.
In general, media player softwares utilizes this information
to play the video.
However, how the playback relies on the header information is
highly variate among softwares.
Also, these softwares normally present post-molded information
as video properties, instead of the raw binary values.
Therefore, I wrote a naive extractor of header information
for video format confirmation.
}
\examples{
\dontrun{
hsvGetAviHeader("input.avi", verbose = TRUE)
}

}
\keyword{utilities}
